requirements_docker.txt
# =============================================================================
# requirements_docker.txt — Dépendances pour les conteneurs Docker
# Sans scrapy-playwright (le scraping se fait en local, pas dans Docker)
# =============================================================================

# ── Nettoyage & Analyse ───────────────────────────────────────────────────────
pandas>=2.2.0
numpy>=1.26.0

# ── Base de données ───────────────────────────────────────────────────────────
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0

# ── API REST ──────────────────────────────────────────────────────────────────
flask>=3.0.0
flask-sqlalchemy>=3.1.0
flask-migrate>=4.0.0
flasgger>=0.9.7.1

# ── Tâches asynchrones ────────────────────────────────────────────────────────
celery>=5.3.0
celery[redis]>=5.3.0
redis>=5.0.0

# ── Monitoring ────────────────────────────────────────────────────────────────
prometheus-flask-exporter>=0.23.0

# ── Utilitaires ───────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
requests>=2.31.0